PID Tuning, CANcoder fix, Deadzone adjustments - Tuned PID for turning only (PID for drive is not being used yet) - Updated Absolute Encoder Values - Realized that the absolute encoder was connected via daisy chains, so swapped to the CANcoder clas instead of the AbsoluteEncoder class - Changed the deadzone for the joystick
Summary of result: PID works well, though does seem slower than the past. This may be because we are only using P and not I and D, affecting the max speed it can turn at.
In addition, the wheels do sound squeaky, so perhaps mechanical may need to add grease.
Also, the wheels would always turn a certain direction after release. We beleive ths is due to the joystick being offset, so we increased the deadband, but have not test.
As for the angle, the wheels do seem off and not straight, so might need to recheck the encoder value. Plus, the rotation joystick does not work so need to fix that later