Changed Presets and Wheel Rotation Settings Got rid of presets 2 and 3 and condensed them into 1. Now the robot can move in all 4 directions, point wheels in 4 directions, rotate wheels, and reset odometry without having to change the controller
Also realized that wheel rotation is based of radians and not degree, so changed presets to reflect this.
In addition, when holding down the button to rotate wheels, it would eventually stop. This is most likely due to the variables not changing.
So to fix this, used DoubleSuppliers to get real data as opposed to the static primitve data that would never update.